move duplicate table functions to separate service
