+++ src/app/analyzer-table/analyzer-table.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, OnChanges, Input, Output, ViewChildren, EventEmitter, AfterViewChecked } from '@angular/core';
 import { AnalyzerService } from '../analyzer.service';
 import { SeriesHelperService } from '../series-helper.service';
+import { TableHelperService } from '../table-helper.service';
 import 'jquery';
 declare var $: any;
 
@@ -23,7 +24,7 @@ export class AnalyzerTableComponent implements OnInit {
   private tableWidths = [];
   private tableDates;
 
-  constructor(private _analyzer: AnalyzerService, private _series: SeriesHelperService) { }
+  constructor(private _analyzer: AnalyzerService, private _series: SeriesHelperService, private _table: TableHelperService) { }
 
   ngOnInit() {
   }
@@ -60,87 +61,22 @@ export class AnalyzerTableComponent implements OnInit {
   ngAfterViewChecked() {
     // Check height of content and scroll tables to the right
     // If true, height is changing, i.e. content still loading
-    if (this.checkContainerHeight()) {
-      this.tableScroll();
+    const heightChange = this._table.checkContainerHeight(this.previousHeight);
+    if (heightChange) {
+      // On load, table scrollbars should start at the right -- showing most recent data
+      return this._table.tableScroll(this.tableEl);
     }
 
     // Scroll tables to the right when table widths changes, i.e. changing frequency from A to Q | M
-    this.checkTableWidth();
-  }
-
-  checkContainerHeight() {
-    const contianer = $('.multi-series-container');
-    const heightDiff = (this.previousHeight !== contianer.height());
-    this.previousHeight = contianer.height();
-    return heightDiff;
-  }
-
-  checkTableWidth() {
-    const tables = $('.table');
-    const tableWidths = this.tableWidths;
-    if (tables) {
-      tables.each(function (index) {
-        const widthDiff = (tableWidths[index] !== tables[index].scrollWidth);
-        if (widthDiff) {
-          tables[index].scrollLeft = tables[index].scrollWidth;
-        }
-        tableWidths[index] = tables[index].scrollWidth;
-      });
-    }
+    return this._table.checkTableWidth(this.tableWidths);
   }
 
   showPopover(seriesInfo) {
-    $('[data-toggle="tooltip"]').tooltip('hide');
-    const popover = $('#' + seriesInfo.id).popover({
-      trigger: 'manual',
-      placement: function (popoverEl, el) {
-        // popoverEl = popover DOM element
-        // el = DOM element that triggers popover
-        let position = 'top';
-        const elOffset = $(el).offset().top;
-        if (elOffset <= 150) {
-          position = 'bottom';
-        }
-        return position;
-      },
-      html: true,
-      title: function () {
-        let title = seriesInfo.title;
-        title += seriesInfo.unitsLabel ? ' (' + seriesInfo.unitsLabel + ')' : ' (' + seriesInfo.unitsLabelShort + ')';
-        return title;
-      },
-      content: function () {
-        let info = '';
-        if (seriesInfo.seasonalAdjustment === 'seasonally_adjusted') {
-          info += 'Seasonally Adjusted<br>';
-        }
-        if (seriesInfo.sourceDescription) {
-          info += 'Source: ' + seriesInfo.sourceDescription + '<br>';
-        }
-        if (seriesInfo.sourceLink) {
-          info += '<a target="_blank" href="' + seriesInfo.sourceLink + '">' + seriesInfo.sourceLink + '</a><br>';
-        }
-        if (seriesInfo.sourceDetails) {
-          info += seriesInfo.sourceDetails;
-        }
-        return info;
-      }
-    }).on('show.bs.popover', function (e) {
-      // Display only one popover at a time
-      $('.popover').not(e.target).popover('dispose');
-      setTimeout(() => {
-        // Close popover on next click (source link in popover is still clickable)
-        $('body').one('click', function () {
-          popover.popover('dispose');
-        });
-      }, 1);
-    });
-    popover.popover('toggle');
+    return this._table.showPopover(seriesInfo);
   }
 
   hideInfo(seriesId) {
-    $('[data-toggle="tooltip"]').tooltip('hide');
-    $('.popover').popover('dispose');
+    return this._table.hideInfo(seriesId);
   }
 
 
@@ -160,16 +96,4 @@ export class AnalyzerTableComponent implements OnInit {
   updateChart(series) {
     this.updateChartSeries.emit(series)
   }
-
-  // On load, table scrollbars should start at the right -- showing most recent data
-  tableScroll(): void {
-    try {
-      this.tableEl._results.forEach((el) => {
-        el.nativeElement.scrollLeft = el.nativeElement.scrollWidth;
-      });
-    } catch (err) {
-      console.log(err);
-    }
-  }
-
 }
