+++ src/app/app.module.ts
@@ -34,6 +34,7 @@ import { HelperService } from './helper.service';
 import { GoogleAnalyticsEventsService } from './google-analytics-events.service';
 import { HelpService } from './help.service';
 import { AnalyzerService } from './analyzer.service';
+import { TableHelperService } from './table-helper.service';
 import { LandingPageComponent } from './landing-page/landing-page.component';
 import { UheroHelpComponent } from './uhero-help/uhero-help.component';
 
@@ -62,6 +63,7 @@ import { UheroHelpComponent } from './uhero-help/uhero-help.component';
     SeriesHelperService,
     HelperService,
     HelpService,
+    TableHelperService,
     GoogleAnalyticsEventsService,
     AnalyzerService,
     Title,
