+++ src/app/category-table/category-table.component.ts
@@ -4,6 +4,7 @@ import { AnalyzerService } from '../analyzer.service';
 import { UheroApiService } from '../uhero-api.service';
 import { GoogleAnalyticsEventsService } from '../google-analytics-events.service';
 import { HelperService } from '../helper.service';
+import { TableHelperService } from '../table-helper.service';
 import { Frequency } from '../frequency';
 import { Geography } from '../geography';
 import 'jquery';
@@ -37,6 +38,7 @@ export class CategoryTableComponent implements AfterViewChecked, OnChanges {
   constructor(
     @Inject('defaultRange') private defaultRange,
     private _uheroAPIService: UheroApiService,
+    private _table: TableHelperService,
     private _helper: HelperService,
     private route: ActivatedRoute,
     private _router: Router,
@@ -79,33 +81,14 @@ export class CategoryTableComponent implements AfterViewChecked, OnChanges {
   ngAfterViewChecked() {
     // Check height of content and scroll tables to the right
     // If true, height is changing, i.e. content still loading
-    if (this.checkContainerHeight()) {
-      this.tableScroll();
+    const heightChange = this._table.checkContainerHeight(this.previousHeight);    
+    if (heightChange) {
+      // On load, table scrollbars should start at the right -- showing most recent data
+      return this._table.tableScroll(this.tableEl);
     }
 
     // Scroll tables to the right when table widths changes, i.e. changing frequency from A to Q | M
-    this.checkTableWidth();
-  }
-
-  checkContainerHeight() {
-    const contianer = $('.multi-series-container');
-    const heightDiff = (this.previousHeight !== contianer.height());
-    this.previousHeight = contianer.height();
-    return heightDiff;
-  }
-
-  checkTableWidth() {
-    const tables = $('.table');
-    const tableWidths = this.tableWidths;
-    if (tables) {
-      tables.each(function (index) {
-        const widthDiff = (tableWidths[index] !== tables[index].scrollWidth);
-        if (widthDiff) {
-          tables[index].scrollLeft = tables[index].scrollWidth;
-        }
-        tableWidths[index] = tables[index].scrollWidth;
-      });
-    }
+    return this._table.checkTableWidth(this.tableWidths);
   }
 
   showTooltip() {
@@ -113,73 +96,16 @@ export class CategoryTableComponent implements AfterViewChecked, OnChanges {
   }
 
   hideInfo(seriesId) {
-    $('[data-toggle="tooltip"]').tooltip('hide');
-    $('.popover').popover('dispose');
     this.submitGAEvent(seriesId);
+    return this._table.hideInfo(seriesId);
   }
 
-  showPopover(subcatIndex, seriesInfo) {
-    $('[data-toggle="tooltip"]').tooltip('hide');
-    const popover = $('#' + subcatIndex + seriesInfo.id).popover({
-      trigger: 'manual',
-      placement: function (popoverEl, el) {
-        // popoverEl = popover DOM element
-        // el = DOM element that triggers popover
-        let position = 'top';
-        const elOffset = $(el).offset().top;
-        if (elOffset <= 150) {
-          position = 'bottom';
-        }
-        return position;
-      },
-      html: true,
-      title: function () {
-        let title = seriesInfo.title;
-        title += seriesInfo.unitsLabel ? ' (' + seriesInfo.unitsLabel + ')' : ' (' + seriesInfo.unitsLabelShort + ')';
-        return title;
-      },
-      content: function () {
-        let info = '';
-        if (seriesInfo.seasonalAdjustment === 'seasonally_adjusted') {
-          info += 'Seasonally Adjusted<br>';
-        }
-        if (seriesInfo.sourceDescription) {
-          info += 'Source: ' + seriesInfo.sourceDescription + '<br>';
-        }
-        if (seriesInfo.sourceLink) {
-          info += '<a target="_blank" href="' + seriesInfo.sourceLink + '">' + seriesInfo.sourceLink + '</a><br>';
-        }
-        if (seriesInfo.sourceDetails) {
-          info += seriesInfo.sourceDetails;
-        }
-        return info;
-      }
-    }).on('show.bs.popover', function (e) {
-      // Display only one popover at a time
-      $('.popover').not(e.target).popover('dispose');
-      setTimeout(() => {
-        // Close popover on next click (source link in popover is still clickable)
-        $('body').one('click', function () {
-          popover.popover('dispose');
-        });
-      }, 1);
-    });
-    popover.popover('toggle');
+  showPopover(seriesInfo, subcatIndex) {
+    return this._table.showPopover(seriesInfo, subcatIndex);
   }
 
   updateAnalyze(seriesInfo, tableData, chartData) {
     this._analyzer.updateAnalyzer(seriesInfo, tableData, chartData);
-  }
-
-  // On load, table scrollbars should start at the right -- showing most recent data
-  tableScroll(): void {
-    try {
-      this.tableEl._results.forEach((el) => {
-        el.nativeElement.scrollLeft = el.nativeElement.scrollWidth;
-      });
-    } catch (err) {
-      console.log(err);
-    }
   }
 
   // Google Analytics: Track clicking on series
